+++ CHANGELOG.md
@@ -1,3 +1,75 @@
+## [1.10.2-beta.12](https://github.com/analogjs/analog/compare/v1.10.2-beta.11...v1.10.2-beta.12) (2024-12-16)
+
+### Bug Fixes
+
+- **vite-plugin-angular:** properly transform files during testing with watch mode ([#1503](https://github.com/analogjs/analog/issues/1503)) ([6d9179f](https://github.com/analogjs/analog/commit/6d9179f5e624d9ff65c8f1069424b07903cd9865))
+
+## [1.10.2-beta.11](https://github.com/analogjs/analog/compare/v1.10.2-beta.10...v1.10.2-beta.11) (2024-12-15)
+
+### Bug Fixes
+
+- **vite-plugin-angular:** only apply sourcemap transform to TS files during testing ([#1500](https://github.com/analogjs/analog/issues/1500)) ([0b56d46](https://github.com/analogjs/analog/commit/0b56d463ac75f3d917919e2f658f171a68b5cd26))
+
+## [1.10.2-beta.10](https://github.com/analogjs/analog/compare/v1.10.2-beta.9...v1.10.2-beta.10) (2024-12-14)
+
+### Bug Fixes
+
+- **vite-plugin-angular:** process destructured objects ([#1499](https://github.com/analogjs/analog/issues/1499)) ([cba76da](https://github.com/analogjs/analog/commit/cba76da9156f2d1c8b4d50d104e057ff5da78c9d))
+
+## [1.10.2-beta.9](https://github.com/analogjs/analog/compare/v1.10.2-beta.8...v1.10.2-beta.9) (2024-12-14)
+
+### Bug Fixes
+
+- **vite-plugin-angular:** process call expr and property access expr property ([#1498](https://github.com/analogjs/analog/issues/1498)) ([cf7db8f](https://github.com/analogjs/analog/commit/cf7db8f05902fa138c8f1e0042c01430e5ef7ab9))
+
+## [1.10.2-beta.8](https://github.com/analogjs/analog/compare/v1.10.2-beta.7...v1.10.2-beta.8) (2024-12-13)
+
+### Bug Fixes
+
+- **vite-plugin-angular:** fix Angular v18/v19 support for ɵPendingTasks ([#1497](https://github.com/analogjs/analog/issues/1497)) ([3215ab4](https://github.com/analogjs/analog/commit/3215ab4e0dfb67d694d660ce21087d182bcb8be5))
+
+## [1.10.2-beta.7](https://github.com/analogjs/analog/compare/v1.10.2-beta.6...v1.10.2-beta.7) (2024-12-12)
+
+### Bug Fixes
+
+- **vite-plugin-angular:** return undefined when not resolving id ([#1495](https://github.com/analogjs/analog/issues/1495)) ([ac6a854](https://github.com/analogjs/analog/commit/ac6a854bcc921d9e4c77383f99bfd63186dce582))
+
+## [1.10.2-beta.6](https://github.com/analogjs/analog/compare/v1.10.2-beta.5...v1.10.2-beta.6) (2024-12-11)
+
+### Bug Fixes
+
+- exclude astro-app from builds ([ebabc90](https://github.com/analogjs/analog/commit/ebabc90b486169796d45f71e1c80d59a1d3fe584))
+
+## [1.10.2-beta.5](https://github.com/analogjs/analog/compare/v1.10.2-beta.4...v1.10.2-beta.5) (2024-12-11)
+
+### Bug Fixes
+
+- **vite-plugin-angular:** resolve JIT external template/style files ([#1493](https://github.com/analogjs/analog/issues/1493)) ([54e9488](https://github.com/analogjs/analog/commit/54e9488f222f56fd055d8b837459eb668c94d0a3))
+
+## [1.10.2-beta.4](https://github.com/analogjs/analog/compare/v1.10.2-beta.3...v1.10.2-beta.4) (2024-12-10)
+
+### Bug Fixes
+
+- **vite-plugin-angular:** skip re-transform in JIT mode ([#1489](https://github.com/analogjs/analog/issues/1489)) ([30533c6](https://github.com/analogjs/analog/commit/30533c6f0deaa3457847a65b170d3938c15ee087))
+
+## [1.10.2-beta.3](https://github.com/analogjs/analog/compare/v1.10.2-beta.2...v1.10.2-beta.3) (2024-12-10)
+
+### Bug Fixes
+
+- **create-analog:** remove override for @nx/angular package ([#1487](https://github.com/analogjs/analog/issues/1487)) ([cb3d70b](https://github.com/analogjs/analog/commit/cb3d70bc1a5a2451c29f43e04d98e29826a07f40))
+
+## [1.10.2-beta.2](https://github.com/analogjs/analog/compare/v1.10.2-beta.1...v1.10.2-beta.2) (2024-12-08)
+
+### Bug Fixes
+
+- **nx-plugin:** update Nx preset to 20.x and Angular 19.x ([#1484](https://github.com/analogjs/analog/issues/1484)) ([7d8ee23](https://github.com/analogjs/analog/commit/7d8ee23498e483a5026a00c819ee05de8e62bc23))
+
+## [1.10.2-beta.1](https://github.com/analogjs/analog/compare/v1.10.1...v1.10.2-beta.1) (2024-12-07)
+
+### Bug Fixes
+
+- **create-analog:** set root property to empty on angular.json ([#1481](https://github.com/analogjs/analog/issues/1481)) ([d9cdbe2](https://github.com/analogjs/analog/commit/d9cdbe20b82145bc24662ae1bdbd5d86b4f202e2))
+
 ## [1.10.1](https://github.com/analogjs/analog/compare/v1.10.0...v1.10.1) (2024-12-04)
 
 ### Bug Fixes